# openapi_spec.yml
openapi: 3.1.0
info:
  title: Advanced Real-Time Discussion Forum API
  description: Backend API for the discussion forum supporting JWT auth, RBAC, threads,
    nested comments, likes, notifications, moderation, and websocket updates.
  version: 1.0.0
servers:
- url: http://localhost:8000
  description: Local backend
- url: http://localhost
  description: Nginx reverse proxy
paths:
...
  /api/v1/threads:
    post:
      tags:
      - Threads
      summary: Create Thread
      operationId: create_thread_api_v1_threads_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThreadCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Threads
      summary: List Threads
      operationId: list_threads_api_v1_threads_get
      security:
      - HTTPBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadListResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/comments:
    post:
      tags:
      - Comments
      summary: Create Comment
      operationId: create_comment_api_v1_comments_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /api/v1/notifications:
    get:
      tags:
      - Notifications
      summary: Get Notifications
      operationId: get_notifications_api_v1_notifications_get
      security:
      - HTTPBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationListResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/notifications/{notification_id}/read:
    patch:
      tags:
      - Notifications
      summary: Mark Notification Read
      operationId: mark_notification_read_api_v1_notifications__notification_id__read_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: notification_id
        in: path
        required: true
        schema:
          type: integer
          title: Notification Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/notifications/read-all:
    patch:
      tags:
      - Notifications
      summary: Mark All Notifications Read
      operationId: mark_all_notifications_read_api_v1_notifications_read_all_patch
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationMarkAllReadResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
      - HTTPBearer: []
  /api/v1/notifications/unread-count:
    get:
      tags:
      - Notifications
      summary: Get Unread Count
      operationId: get_unread_count_api_v1_notifications_unread_count_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationUnreadCountResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
      - HTTPBearer: []
  /api/v1/moderation:
    post:
      tags:
      - Moderation
      summary: Create Review
      operationId: create_review_api_v1_moderation_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModerationCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModerationResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/moderation/report:
    post:
      tags:
      - Moderation
      summary: Report Content
      operationId: report_content_api_v1_moderation_report_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModerationResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/moderation/pending:
    get:
      tags:
      - Moderation
      summary: Get Pending Reviews
      operationId: get_pending_reviews_api_v1_moderation_pending_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ModerationResponse'
                type: array
                title: Response Get Pending Reviews Api V1 Moderation Pending Get
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
      - HTTPBearer: []
  /api/v1/moderation/completed:
    get:
      tags:
      - Moderation
      summary: Get Completed Reviews
      operationId: get_completed_reviews_api_v1_moderation_completed_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ModerationResponse'
                type: array
                title: Response Get Completed Reviews Api V1 Moderation Completed
                  Get
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
      - HTTPBearer: []
  /api/v1/moderation/{review_id}:
    put:
      tags:
      - Moderation
      summary: Update Review
      operationId: update_review_api_v1_moderation__review_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: review_id
        in: path
        required: true
        schema:
          type: integer
          title: Review Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModerationUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModerationResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/moderation/{review_id}/action:
    post:
      tags:
      - Moderation
      summary: Take Action On Review
      operationId: take_action_on_review_api_v1_moderation__review_id__action_post
      security:
      - HTTPBearer: []
      parameters:
      - name: review_id
        in: path
        required: true
        schema:
          type: integer
          title: Review Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModerationUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModerationResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/search/threads:
    get:
      tags:
      - Search
      summary: Search Threads
      operationId: search_threads_api_v1_search_threads_get
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          title: Q
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Size
      - name: searchIn
        in: query
        required: false
        schema:
          type: string
          pattern: ^(all|title|content|tags)$
          default: all
          title: Searchin
      - name: sortBy
        in: query
        required: false
        schema:
          type: string
          pattern: ^(relevance|recent|popular)$
          default: relevance
          title: Sortby
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadSearchResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/search/comments:
    get:
      tags:
      - Search
      summary: Search Comments
      operationId: search_comments_api_v1_search_comments_get
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          title: Q
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentSearchResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CommentCreate:
      properties:
        thread_id:
          type: integer
          title: Thread Id
        content:
          type: string
          title: Content
        parent_comment_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Parent Comment Id
      type: object
      required:
      - thread_id
      - content
      title: CommentCreate
    CommentResponse:
      properties:
        id:
          type: integer
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        content:
          type: string
          title: Content
        thread_id:
          type: integer
          title: Thread Id
        author_id:
          type: integer
          title: Author Id
        author:
          anyOf:
          - $ref: '#/components/schemas/ThreadAuthor'
          - type: 'null'
        parent_comment_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Parent Comment Id
        like_count:
          type: integer
          title: Like Count
          default: 0
        user_has_liked:
          type: boolean
          title: User Has Liked
          default: false
        is_deleted:
          type: boolean
          title: Is Deleted
      type: object
      required:
      - id
      - created_at
      - updated_at
      - content
      - thread_id
      - author_id
      - parent_comment_id
      - is_deleted
      title: CommentResponse
    CommentSearchResponse:
      properties:
        results:
          items:
            $ref: '#/components/schemas/CommentResponse'
          type: array
          title: Results
        total:
          type: integer
          title: Total
      type: object
      required:
      - results
      - total
      title: CommentSearchResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ModerationCreate:
      properties:
        content_type:
          type: string
          title: Content Type
        thread_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Thread Id
        comment_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Comment Id
      type: object
      required:
      - content_type
      title: ModerationCreate
    ModerationResponse:
      properties:
        id:
          type: integer
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        content_type:
          type: string
          title: Content Type
        thread_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Thread Id
        comment_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Comment Id
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
        reviewer_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Reviewer Id
        status:
          type: string
          title: Status
        action_taken:
          anyOf:
          - type: string
          - type: 'null'
          title: Action Taken
      type: object
      required:
      - id
      - created_at
      - updated_at
      - content_type
      - thread_id
      - comment_id
      - reason
      - reviewer_id
      - status
      - action_taken
      title: ModerationResponse
    ModerationUpdate:
      properties:
        status:
          type: string
          title: Status
        action_taken:
          anyOf:
          - type: string
          - type: 'null'
          title: Action Taken
      type: object
      required:
      - status
      title: ModerationUpdate
    NotificationListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/NotificationResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
      type: object
      required:
      - items
      - total
      - page
      - size
      title: NotificationListResponse
    NotificationMarkAllReadResponse:
      properties:
        updated:
          type: integer
          minimum: 0.0
          title: Updated
      type: object
      required:
      - updated
      title: NotificationMarkAllReadResponse
    NotificationResponse:
      properties:
        id:
          type: integer
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        user_id:
          type: integer
          title: User Id
        actor_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Actor Id
        type:
          type: string
          title: Type
        title:
          type: string
          title: Title
        message:
          type: string
          title: Message
        entity_type:
          type: string
          title: Entity Type
        entity_id:
          type: integer
          title: Entity Id
        is_read:
          type: boolean
          title: Is Read
      type: object
      required:
      - id
      - created_at
      - updated_at
      - user_id
      - actor_id
      - type
      - title
      - message
      - entity_type
      - entity_id
      - is_read
      title: NotificationResponse
    NotificationUnreadCountResponse:
      properties:
        unread_count:
          type: integer
          title: Unread Count
      type: object
      required:
      - unread_count
      title: NotificationUnreadCountResponse
...
    ReportCreate:
      properties:
        content_type:
          type: string
          title: Content Type
        thread_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Thread Id
        comment_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Comment Id
        reason:
          type: string
          title: Reason
      type: object
      required:
      - content_type
      - reason
      title: ReportCreate
...
    ThreadAuthor:
      properties:
        id:
          type: integer
          title: Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        email:
          type: string
          title: Email
        avatar_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Avatar Url
      type: object
      required:
      - id
      - email
      title: ThreadAuthor
    ThreadCreate:
      properties:
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        tags:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Tags
      type: object
      required:
      - title
      - description
      title: ThreadCreate
    ThreadListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ThreadResponse'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
        pages:
          type: integer
          title: Pages
      type: object
      required:
      - items
      - total
      - page
      - size
      - pages
      title: ThreadListResponse
    ThreadResponse:
      properties:
        id:
          type: integer
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        tags:
          items:
            type: string
          type: array
          title: Tags
        author_id:
          type: integer
          title: Author Id
        author:
          anyOf:
          - $ref: '#/components/schemas/ThreadAuthor'
          - type: 'null'
        comment_count:
          type: integer
          title: Comment Count
          default: 0
        like_count:
          type: integer
          title: Like Count
          default: 0
        user_has_liked:
          type: boolean
          title: User Has Liked
          default: false
        is_deleted:
          type: boolean
          title: Is Deleted
      type: object
      required:
      - id
      - created_at
      - updated_at
      - title
      - description
      - author_id
      - is_deleted
      title: ThreadResponse
    ThreadSearchResponse:
      properties:
        results:
          items:
            $ref: '#/components/schemas/ThreadResponse'
          type: array
          title: Results
        total:
          type: integer
          title: Total
      type: object
      required:
      - results
      - total
      title: ThreadSearchResponse
      description: Search result wrapper.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...